+++ discollection-cli/package.json
@@ -1,0 +1,42 @@
+{
+  "author": "",
+  "bin": {
+    "discollection": "./dist/bin/index.js"
+  },
+  "dependencies": {
+    "cac": "^7.0.0",
+    "chalk": "^5.3.0",
+    "chalk-table": "^1.0.2",
+    "inquirer": "^10.1.2",
+    "invariant": "^2.2.4",
+    "ky": "^1.5.0",
+    "lodash": "^4.17.21",
+    "tslog": "^4.9.3"
+  },
+  "description": "",
+  "devDependencies": {
+    "@tsconfig/node20": "^20.1.4",
+    "@types/invariant": "^2.2.37",
+    "@types/lodash": "^4.17.7",
+    "@types/node": "^22.0.0",
+    "prettier": "3.3.3",
+    "tsup": "^8.2.4",
+    "tsx": "^4.16.5",
+    "typescript": "^5.5.4"
+  },
+  "exports": "./dist/bin/index.js",
+  "keywords": [],
+  "license": "ISC",
+  "name": "discollection",
+  "packageManager": "pnpm@9.6.0+sha512.38dc6fba8dba35b39340b9700112c2fe1e12f10b17134715a4aa98ccf7bb035e76fd981cf0bb384dfa98f8d6af5481c2bef2f4266a24bfa20c34eb7147ce0b5e",
+  "scripts": {
+    "build": "tsup",
+    "dev": "tsx src/cli.ts",
+    "format": "prettier . --check",
+    "format:fix": "prettier . --write",
+    "start": "node dist/bin/index.js",
+    "start:node": "node dist/bin/index.js"
+  },
+  "type": "module",
+  "version": "1.0.0"
+}
