--- prisma/schema.prisma
@@ -1,30 +1,0 @@
-generator client {
-  provider = "prisma-client-js"
-}
-
-datasource db {
-  provider = "sqlite"
-  url      = env("DATABASE_URL")
-}
-
-model Release {
-  id     Int            @id @default(autoincrement())
-  title  String
-  artist String?
-  genres ReleaseGenre[]
-  styles ReleaseStyle[]
-}
-
-model ReleaseGenre {
-  id        Int     @id @default(autoincrement())
-  releaseId Int
-  genre     String
-  release   Release @relation(fields: [releaseId], references: [id])
-}
-
-model ReleaseStyle {
-  id        Int     @id @default(autoincrement())
-  releaseId Int
-  style     String
-  release   Release @relation(fields: [releaseId], references: [id])
-}
