--- src/utilities/formatter.ts
@@ -1,8 +1,0 @@
-import { Release } from "@prisma/client";
-
-export function toReleaseObject(release: Release) {
-  return {
-    title: release.title,
-    artist: release.artist,
-  };
-}
