--- tsup.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-  entry: ["bin/run.ts"],
-  splitting: false,
-  sourcemap: false,
-  clean: true,
-});
